Implement load operation (#73)
